Rooms implemented with icons

+++ app/src/main/java/com/korinek/indoorlocalizatorapp/ui/building/BuildingFragment.java
@@ -1,11 +1,16 @@
 package com.korinek.indoorlocalizatorapp.ui.building;
 
+import android.content.DialogInterface;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.AdapterView;
+import android.widget.ArrayAdapter;
 import android.widget.Button;
 import android.widget.EditText;
+import android.widget.ImageView;
+import android.widget.Spinner;
 import android.widget.TextView;
 import android.widget.Toast;
 
@@ -13,8 +18,7 @@ import androidx.annotation.NonNull;
 import androidx.appcompat.app.AlertDialog;
 import androidx.fragment.app.Fragment;
 import androidx.lifecycle.ViewModelProvider;
-import androidx.recyclerview.widget.ItemTouchHelper;
-import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.GridLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.korinek.indoorlocalizatorapp.R;
@@ -27,6 +31,7 @@ public class BuildingFragment extends Fragment {
 
     private FragmentBuildingBinding binding;
     BuildingViewModel buildingViewModel;
+    private int selectedIcon;
 
     public View onCreateView(@NonNull LayoutInflater inflater,
                              ViewGroup container, Bundle savedInstanceState) {
@@ -55,12 +60,40 @@ public class BuildingFragment extends Fragment {
             @Override
             public void onRoomClick(Room room) {
                 System.out.println("Room click: " + room.getName());
+                Toast.makeText(requireContext(), "Click: " + room.getName(), Toast.LENGTH_LONG).show();
+                // TODO - implement what to do
             }
 
             @Override
-            public void onRoomSwiped(Room room) {
-                System.out.println("Room swiped: " + room.getName());
-                buildingViewModel.deleteRoom(room);
+            public void onRoomCalibrate(Room room) {
+                System.out.println("Room calibrate click: " + room.getName());
+                Toast.makeText(requireContext(), "Click: " + room.getName() + " - Kalibrovat", Toast.LENGTH_LONG).show();
+                // TODO - implement what to do
+            }
+
+            @Override
+            public void onRoomEdit(Room room) {
+                System.out.println("Room edit click: " + room.getName());
+                Toast.makeText(requireContext(), "Click: " + room.getName() + " - Upravit", Toast.LENGTH_LONG).show();
+                // TODO - implement what to do
+            }
+
+            @Override
+            public void onRoomDelete(Room room) {
+                System.out.println("Room delete click: " + room.getName());
+
+                new AlertDialog.Builder(requireContext())
+                        .setTitle("Odstranění místnosti")
+                        .setMessage(String.format("Opravdu chcete odstranit místnost %s?", room.getName()))
+                        //.setMessage(message)
+                        .setPositiveButton("Odstranit", (dialog, which) -> {
+                            buildingViewModel.deleteRoom(room);
+                        })
+                        .setNegativeButton("Zrušit", (dialog, which) -> {
+                            dialog.dismiss();
+                        })
+                        .setOnCancelListener(DialogInterface::dismiss)
+                        .show();
             }
         });
 
@@ -74,38 +107,7 @@ public class BuildingFragment extends Fragment {
         });
 
         roomsRecyclerView.setAdapter(roomAdapter);
-        roomsRecyclerView.setLayoutManager(new LinearLayoutManager(requireContext()));
-
-        new ItemTouchHelper(new ItemTouchHelper.SimpleCallback(0, ItemTouchHelper.LEFT) {
-            @Override
-            public boolean onMove(@NonNull RecyclerView recyclerView, @NonNull RecyclerView.ViewHolder viewHolder, @NonNull RecyclerView.ViewHolder target) {
-                return false;
-            }
-
-            @Override
-            public void onSwiped(@NonNull RecyclerView.ViewHolder viewHolder, int direction) {
-                int position = viewHolder.getAdapterPosition();
-                Room room = roomAdapter.getRoomAt(position);
-                //String message = getString(R.string.delete_building_message, building.getName());
-
-                new AlertDialog.Builder(requireContext())
-                        .setTitle("Odstranění místnosti")
-                        .setMessage(String.format("Opravdu chcete odstranit místnost %s?", room.getName()))
-                        //.setMessage(message)
-                        .setPositiveButton("Odstranit", (dialog, which) -> {
-                            roomAdapter.getListener().onRoomSwiped(room);
-                        })
-                        .setNegativeButton("Zrušit", (dialog, which) -> {
-                            // return the room back
-                            roomAdapter.notifyItemChanged(position);
-                        })
-                        .setOnCancelListener(dialog -> {
-                            // return the room back
-                            roomAdapter.notifyItemChanged(position);
-                        })
-                        .show();
-            }
-        }).attachToRecyclerView(roomsRecyclerView);
+        roomsRecyclerView.setLayoutManager(new GridLayoutManager(requireContext(), 2));
 
         buildingViewModel.getIsBuildingSelected().observe(getViewLifecycleOwner(), isBuildingSelected -> {
             textBuildingNotSet.setVisibility(isBuildingSelected ? View.GONE : View.VISIBLE);
@@ -128,6 +130,9 @@ public class BuildingFragment extends Fragment {
         View dialogView = LayoutInflater.from(requireContext()).inflate(R.layout.dialog_add_room, null);
 
         EditText roomNameInput = dialogView.findViewById(R.id.room_name_input);
+        Spinner iconSelector = dialogView.findViewById(R.id.room_icon_selector);
+
+        initializeIconSelector(iconSelector);
 
         new AlertDialog.Builder(requireContext())
                 .setTitle("Přidat novou místnost")
@@ -135,9 +140,9 @@ public class BuildingFragment extends Fragment {
                 .setPositiveButton("Přidat", (dialog, which) -> {
                     String roomName = roomNameInput.getText().toString().trim();
 
-                    if (!roomName.isEmpty()) {
+                    if (!roomName.isEmpty() && selectedIcon != 0) {
                         int buildingId = buildingViewModel.getSelectedBuilding().getId();
-                        Room room = new Room(roomName, buildingId);
+                        Room room = new Room(roomName, buildingId, selectedIcon);
                         buildingViewModel.insertRoom(room);
                     } else {
                         Toast.makeText(requireContext(), "Chyba: Je třeba zadat název místnosti!", Toast.LENGTH_LONG).show();
@@ -145,6 +150,59 @@ public class BuildingFragment extends Fragment {
                 })
                 .setNegativeButton("Zrušit", (dialog, which) -> dialog.dismiss())
                 .show();
+    }
+
+    private void initializeIconSelector(Spinner iconSelector) {
+        int[] icons = {
+                R.drawable.ic_room_pc_desktop,
+                R.drawable.ic_room_meeting_board,
+                R.drawable.ic_room_floor_plan,
+                R.drawable.ic_room_sofa,
+                R.drawable.ic_room_bed,
+                R.drawable.ic_room_kitchen_faucet,
+                R.drawable.ic_room_fridge,
+                R.drawable.ic_room_shower,
+                R.drawable.ic_room_toilet
+        };
+        ArrayAdapter<Integer> selectorAdapter = new ArrayAdapter<Integer>(requireContext(), android.R.layout.simple_spinner_item) {
+            @NonNull
+            @Override
+            public View getView(int position, View convertView, ViewGroup parent) {
+                if (convertView == null) {
+                    convertView = LayoutInflater.from(getContext()).inflate(R.layout.spinner_room_icon_item, parent, false);
+                }
+                ImageView imageView = convertView.findViewById(R.id.spinner_icon);
+                imageView.setImageResource(icons[position]);
+                return convertView;
+            }
+
+            @Override
+            public View getDropDownView(int position, View convertView, ViewGroup parent) {
+                if (convertView == null) {
+                    convertView = LayoutInflater.from(getContext()).inflate(R.layout.spinner_room_icon_item, parent, false);
+                }
+                ImageView imageView = convertView.findViewById(R.id.spinner_icon);
+                imageView.setImageResource(icons[position]);
+                return convertView;
+            }
+
+            @Override
+            public int getCount() {
+                return icons.length;
+            }
+        };
+        iconSelector.setAdapter(selectorAdapter);
+        iconSelector.setOnItemSelectedListener(new AdapterView.OnItemSelectedListener() {
+            @Override
+            public void onItemSelected(AdapterView<?> parent, View view, int position, long id) {
+                selectedIcon = icons[position];
+            }
+
+            @Override
+            public void onNothingSelected(AdapterView<?> parent) {
+                // Když není nic vybráno
+            }
+        });
     }
 
     @Override
